% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_irace_vizz}
\alias{run_irace_vizz}
\title{Run Irace Studio}
\usage{
run_irace_vizz(port = NULL, ...)
}
\arguments{
\item{port}{A port number that Irace Studio will listen on.

#' @export
run_irace_studio <- function(port = 4350, ...) {
  options(golem.app.prod = TRUE)
  run_app(port, ...)
}}
}
\description{
Run Irace Studio
}
